Reformat AST dump to be valid scad files.
This effectively makes an "inlined" scad file that should behave the same way as
the input file, and as such, when a .scad file is specified as the output, this
will be the default generation.

The annotated variable assignments have been moved to the top of the file, such
that they can be picked up by the customizer.

The motivation for this is to be able to inline libraries that are not included
on Thingiverse.